The customer that was getting that Morph class not found issue was due to the two lines in this file, which prevented him from A) loading Morph on the frontend on a specific page, if Morph was not the default template and B) not load Configurator if Morph was not the default template.
Essentially the customer has Absolute Flush set as the default, as they are working on the live site and thus have Morph assigned to a hidden menu item, that they can use to test. After i made the two changes in Morph cache, everything works as expected (Configurator works properly even if Morph isnt the default, Flush works as the default and Morph works on the page its assigned too.

Please review the changes and let me know if there are any issues.